split each adapt_rk variant into their own .cpp/.h files
