+++ cpp/memilio/math/vadapt_rk_fast.cpp
@@ -17,7 +17,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-#include "memilio/math/adapt_rk.h"
+#include "memilio/math/vadapt_rk_fast.h"
 
 namespace mio
 {
